simple calculator results, remove go-to-bed results component, use @apply css rules
